Fixed PrintSpi messages

+++ src/printspi.cpp
@@ -7,12 +7,12 @@
 using namespace std;
 
 err_t PrintSpi::chipSelect(uint8_t level) {
-    cout << "spi chip select: " << level << endl;
+    cout << "spi chip select: " << static_cast<int>(level) << endl;
     return 0;
 }
 
 err_t PrintSpi::setMode(uint8_t mode) {
-    cout << "spi set mode: " << mode << endl;
+    cout << "spi set mode: " << static_cast<int>(mode) << endl;
     return 0;
 }
 
@@ -31,7 +31,7 @@ err_t PrintSpi::transfer(uint8_t *buf, uint32_t len) {
 }
 
 PrintSpi::PrintSpi(uint8_t index, uint8_t csPin) {
-    cout << "spi " << index << " init, csPin " << csPin << endl;
+    cout << "spi " << static_cast<int>(index) << " init, csPin " << static_cast<int>(csPin) << endl;
 }
 
 PrintSpi::~PrintSpi() {
